+++ projectroles/tests/test_permissions.py
@@ -789,6 +789,7 @@ class TestProjectViews(TestProjectPermissionBase):
             self.assertFalse(data['results'])
 
 
+@override_settings(PROJECTROLES_SITE_MODE=SITE_MODE_TARGET)
 class TestTargetProjectViews(
     RemoteSiteMixin, RemoteProjectMixin, TestProjectPermissionBase
 ):
@@ -820,7 +821,6 @@ class TestTargetProjectViews(
             level=SODAR_CONSTANTS['REMOTE_LEVEL_READ_ROLES'],
         )
 
-    @override_settings(PROJECTROLES_SITE_MODE=SITE_MODE_TARGET)
     def test_update(self):
         """Test access to project updating as target"""
         url = reverse(
@@ -837,7 +837,6 @@ class TestTargetProjectViews(
         ]
         self.assert_redirect(url, bad_users)
 
-    @override_settings(PROJECTROLES_SITE_MODE=SITE_MODE_TARGET)
     def test_create_top_allowed(self):
         """Test access to top level project creation as target"""
         url = reverse('projectroles:create')
@@ -853,7 +852,6 @@ class TestTargetProjectViews(
         self.assert_render200_ok(url, good_users)
         self.assert_redirect(url, bad_users)
 
-    @override_settings(PROJECTROLES_SITE_MODE=SITE_MODE_TARGET)
     def test_create_sub(self):
         """Test access to subproject creation as target"""
         url = reverse(
@@ -870,10 +868,7 @@ class TestTargetProjectViews(
         self.assert_render200_ok(url, good_users)
         self.assert_redirect(url, bad_users)
 
-    @override_settings(
-        PROJECTROLES_SITE_MODE=SITE_MODE_TARGET,
-        PROJECTROLES_TARGET_CREATE=False,
-    )
+    @override_settings(PROJECTROLES_TARGET_CREATE=False)
     def test_create_sub_disallowed(self):
         """Test access to subproject creation with creation disallowed as target"""
         url = reverse(
@@ -890,7 +885,6 @@ class TestTargetProjectViews(
         ]
         self.assert_redirect(url, bad_users)
 
-    @override_settings(PROJECTROLES_SITE_MODE=SITE_MODE_TARGET)
     def test_role_create(self):
         """Test access to role creation as target"""
         url = reverse(
@@ -908,7 +902,6 @@ class TestTargetProjectViews(
         ]
         self.assert_redirect(url, bad_users)
 
-    @override_settings(PROJECTROLES_SITE_MODE=SITE_MODE_TARGET)
     def test_role_update(self):
         """Test access to role updating as target"""
         url = reverse(
@@ -926,7 +919,6 @@ class TestTargetProjectViews(
         ]
         self.assert_redirect(url, bad_users)
 
-    @override_settings(PROJECTROLES_SITE_MODE=SITE_MODE_TARGET)
     def test_role_delete(self):
         """Test access to role deletion as target"""
         url = reverse(
@@ -944,7 +936,6 @@ class TestTargetProjectViews(
         ]
         self.assert_redirect(url, bad_users)
 
-    @override_settings(PROJECTROLES_SITE_MODE=SITE_MODE_TARGET)
     def test_role_update_delegate(self):
         """Test access to delegate role update as target"""
         url = reverse(
@@ -962,7 +953,6 @@ class TestTargetProjectViews(
         ]
         self.assert_redirect(url, bad_users)
 
-    @override_settings(PROJECTROLES_SITE_MODE=SITE_MODE_TARGET)
     def test_role_delete_delegate(self):
         """Test access to role deletion for delegate as target"""
         url = reverse(
@@ -980,7 +970,6 @@ class TestTargetProjectViews(
         ]
         self.assert_redirect(url, bad_users)
 
-    @override_settings(PROJECTROLES_SITE_MODE=SITE_MODE_TARGET)
     def test_role_invite_create(self):
         """Test access to role invite creation as target"""
         url = reverse(
@@ -998,7 +987,6 @@ class TestTargetProjectViews(
         ]
         self.assert_redirect(url, bad_users)
 
-    @override_settings(PROJECTROLES_SITE_MODE=SITE_MODE_TARGET)
     def test_role_invite_list(self):
         """Test access to role invite list as target"""
         url = reverse(
@@ -1013,6 +1001,69 @@ class TestTargetProjectViews(
             self.as_guest.user,
             self.user_no_roles,
         ]
+        self.assert_redirect(url, bad_users)
+
+
+@override_settings(PROJECTROLES_SITE_MODE=SITE_MODE_TARGET)
+class TestRevokedRemoteProject(
+    RemoteSiteMixin, RemoteProjectMixin, TestProjectPermissionBase
+):
+    """Tests for views for a revoked project on a TARGET site"""
+
+    def setUp(self):
+        super().setUp()
+
+        # Create site
+        self.site = self._make_site(
+            name=REMOTE_SITE_NAME,
+            url=REMOTE_SITE_URL,
+            mode=SODAR_CONSTANTS['SITE_MODE_SOURCE'],
+            description='',
+            secret=REMOTE_SITE_SECRET,
+        )
+
+        # Create RemoteProject objects
+        self.remote_category = self._make_remote_project(
+            project_uuid=self.category.sodar_uuid,
+            project=self.category,
+            site=self.site,
+            level=SODAR_CONSTANTS['REMOTE_LEVEL_READ_INFO'],
+        )
+        self.remote_project = self._make_remote_project(
+            project_uuid=self.project.sodar_uuid,
+            project=self.project,
+            site=self.site,
+            level=SODAR_CONSTANTS['REMOTE_LEVEL_REVOKED'],
+        )
+
+    def test_project_details(self):
+        """Test access to REVOKED project detail page as target"""
+        url = reverse(
+            'projectroles:detail', kwargs={'project': self.project.sodar_uuid}
+        )
+        good_users = [self.superuser, self.as_owner.user, self.as_delegate.user]
+        bad_users = [
+            self.anonymous,
+            self.as_contributor.user,
+            self.as_guest.user,
+            self.user_no_roles,
+        ]
+        self.assert_response(url, good_users, 200)
+        self.assert_redirect(url, bad_users)
+
+    def test_role_list(self):
+        """Test access to REVOKED project's role list as target"""
+        url = reverse(
+            'projectroles:roles', kwargs={'project': self.project.sodar_uuid}
+        )
+        good_users = [self.superuser, self.as_owner.user, self.as_delegate.user]
+        bad_users = [
+            self.anonymous,
+            self.as_contributor.user,
+            self.as_guest.user,
+            self.user_no_roles,
+        ]
+        self.assert_response(url, good_users, 200)
         self.assert_redirect(url, bad_users)