+++ projectroles/tests/test_ui.py
@@ -738,10 +738,9 @@ class TestProjectDetail(TestUIBase, RemoteSiteMixin, RemoteProjectMixin):
 
     @override_settings(PROJECTROLES_SITE_MODE=SITE_MODE_TARGET)
     def test_peer_project_target_visible(self):
-        """Checks visibility of peer projects on TARGET site (user_display=False)"""
+        """Test visibility of peer projects on TARGET site (user_display=False)"""
         # There needs to be a source mode remote project as master project
         # otherwise peer project logic wont be reached
-        # TODO: Clean up when implementing #196 (refactoring project_detail.html logic)
         source_site = self._make_site(
             name='Second Remote Site',
             url='second_remote.site',
@@ -750,14 +749,12 @@ class TestProjectDetail(TestUIBase, RemoteSiteMixin, RemoteProjectMixin):
             secret=build_secret(),
             user_display=True,
         )
-
         self._make_remote_project(
             project_uuid=self.project.sodar_uuid,
             site=source_site,
             level=SODAR_CONSTANTS['REMOTE_LEVEL_READ_ROLES'],
             project=self.project,
         )
-
         self.setup_remote_project(site_mode=SITE_MODE_PEER)
 
         url = reverse(
@@ -773,26 +770,24 @@ class TestProjectDetail(TestUIBase, RemoteSiteMixin, RemoteProjectMixin):
 
         for user in users:
             self.login_and_redirect(user, url)
-
-            projects_on_other_sites = self.selenium.find_element_by_id(
-                'sodar-pr-details-card-remote'
+            remote_links = self.selenium.find_elements_by_class_name(
+                'sodar-pr-link-remote'
             )
-
-            details = projects_on_other_sites.find_elements_by_css_selector('a')
-            self.assertEqual(len(details), 2)
-            self.assertEqual(
-                details[0].text, source_site.name + ' (Master Project)'
+            self.assertEqual(len(remote_links), 2)
+            self.assertIn(
+                'sodar-pr-link-remote-master',
+                remote_links[0].get_attribute('class'),
             )
-            self.assertEqual(
-                details[1].text, self.remote_site.name + ' (Peer Project)'
+            self.assertIn(
+                'sodar-pr-link-remote-peer',
+                remote_links[1].get_attribute('class'),
             )
 
     @override_settings(PROJECTROLES_SITE_MODE=SITE_MODE_TARGET)
     def test_peer_project_target_invisible(self):
-        """Checks invisibility of peer projects on TARGET site for different users (user_display=False)"""
+        """Test invisibility of peer projects on TARGET site for users (user_display=False)"""
         # There needs to be a source mode remote project as master project
         # otherwise peer project logic wont be reached
-        # TODO: Clean up when implementing #196 (refactoring project_detail.html logic)
         source_site = self._make_site(
             name='Second Remote Site',
             url='second_remote.site',
@@ -825,29 +820,28 @@ class TestProjectDetail(TestUIBase, RemoteSiteMixin, RemoteProjectMixin):
 
         for user in expected_true:
             self.login_and_redirect(user, url)
-
-            projects_on_other_sites = self.selenium.find_element_by_id(
-                'sodar-pr-details-card-remote'
+            remote_links = self.selenium.find_elements_by_class_name(
+                'sodar-pr-link-remote'
             )
-            details = projects_on_other_sites.find_elements_by_css_selector('a')
-            self.assertEqual(len(details), 2)
-            self.assertEqual(
-                details[0].text, source_site.name + ' (Master Project)'
+            self.assertEqual(len(remote_links), 2)
+            self.assertIn(
+                'sodar-pr-link-remote-master',
+                remote_links[0].get_attribute('class'),
             )
-            self.assertEqual(
-                details[1].text, self.remote_site.name + ' (Peer Project)'
+            self.assertIn(
+                'sodar-pr-link-remote-peer',
+                remote_links[1].get_attribute('class'),
             )
 
         for user in expected_false:
             self.login_and_redirect(user, url)
-
-            projects_on_other_sites = self.selenium.find_element_by_id(
-                'sodar-pr-details-card-remote'
+            remote_links = self.selenium.find_elements_by_class_name(
+                'sodar-pr-link-remote'
             )
-            details = projects_on_other_sites.find_elements_by_css_selector('a')
-            self.assertEqual(len(details), 1)
-            self.assertEqual(
-                details[0].text, source_site.name + ' (Master Project)'
+            self.assertEqual(len(remote_links), 1)
+            self.assertIn(
+                'sodar-pr-link-remote-master',
+                remote_links[0].get_attribute('class'),
             )
 
     @override_settings(PROJECTROLES_SITE_MODE=SITE_MODE_TARGET)
@@ -915,24 +909,17 @@ class TestProjectRoles(RemoteTargetMixin, TestUIBase):
     def test_list_buttons(self):
         """Test visibility of role list button group according to user
         permissions"""
-        users_with_perm = [
-            self.superuser,
-            self.as_owner.user,
-            self.as_delegate.user,
-        ]
-
-        users_without_perm = [self.as_contributor.user, self.as_guest.user]
-
+        good_users = [self.superuser, self.as_owner.user, self.as_delegate.user]
+        bad_users = [self.as_contributor.user, self.as_guest.user]
         url = reverse(
             'projectroles:roles', kwargs={'project': self.project.sodar_uuid}
         )
 
         self.assert_element_exists(
-            users_with_perm, url, 'sodar-pr-btn-perm', True
+            good_users, url, 'sodar-pr-btn-role-op', True
         )
-
         self.assert_element_exists(
-            users_without_perm, url, 'sodar-pr-btn-no-perm', True
+            bad_users, url, 'sodar-pr-btn-role-op', False
         )
 
     @override_settings(PROJECTROLES_SITE_MODE=SITE_MODE_TARGET)
@@ -942,8 +929,7 @@ class TestProjectRoles(RemoteTargetMixin, TestUIBase):
         # Set up site as target
         self._set_up_as_target(projects=[self.category, self.project])
 
-        users = [
-            self.superuser,
+        non_superusers = [
             self.as_owner.user,
             self.as_delegate.user,
             self.as_contributor.user,
@@ -953,7 +939,13 @@ class TestProjectRoles(RemoteTargetMixin, TestUIBase):
             'projectroles:roles', kwargs={'project': self.project.sodar_uuid}
         )
 
-        self.assert_element_exists(users, url, 'sodar-pr-btn-remote', True)
+        self.login_and_redirect(self.superuser, url)
+        btn = self.selenium.find_element_by_id('sodar-pr-btn-role-op')
+        self.assertEqual(btn.is_enabled(), False)
+
+        self.assert_element_exists(
+            non_superusers, url, 'sodar-pr-btn-role-op', False
+        )
 
     def test_role_list_invite_button(self):
         """Test visibility of role invite button according to user